// web/src/gen/model/pipelineInputDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Automation.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { PinCardinality } from "./pinCardinality";
import type { PinPrimitiveType } from "./pinPrimitiveType";

export interface PipelineInputDto {
  id: string;
  key: string;
  label: string;
  type: PinPrimitiveType;
  cardinality: PinCardinality;
  isRequired: boolean;
  /** @nullable */
  defaultValue: string | null;
  order: number;
}